+++ catscripts/Math/Fractions/Fractions.sh
@@ -1,21 +1,30 @@
 #!/bin/bash
 
-egrep -i 'as a fraction\]\]|into a fraction\]\]' newpages.txt > Fractions.txt
-egrep -i 'in simplest form\]\]' newpages.txt >> Fractions.txt
-egrep -i 'in its simplest form\]\]' newpages.txt >> Fractions.txt
-egrep -i 'in lowest terms\]\]|reduced fraction' newpages.txt >> Fractions.txt
-egrep -i 'simplest form of [0-9]{1,}/[0-9]{1,}' newpages.txt >> Fractions.txt
-egrep -i '[0-9]{1,}(/|\\)[0-9]{1,} of [0-9]{1,}' newpages.txt >> Fractions.txt
-egrep -i '[0-9]{1,}(/|\\)[0-9]{1,}(| )(plus|minus|times|divided(| )by)(| )[0-9]{1,}(/|\\)[0-9]{1,}' newpages.txt >> Fractions.txt
-egrep -i '[0-9]{1,}(/|\\)[0-9]{1,}(| )(plus|minus|times|divided(| )by)(| )[0-9]{1.} [0-9]{1,}(/|\\)[0-9]{1,}' newpages.txt >> Fractions.txt
+KEYWORDS_FRACTIONS="'as(| )a(| )fraction$|into(| )a(| )fraction$|in(| )simplest(| )form$|in(| )its(| )simplest(| )form$|in(| )lowest(| )terms$|reduced(| )fraction|simplest(| )form(| )of(| )[0-9]{1,}/[0-9]{1,}|[0-9]{1,}(/|\\)[0-9]{1,}(| )of(| )[0-9]{1,}|[0-9]{1,}(/|\\)[0-9]{1,}(| )(plus|minus|times|divided(| )by)(| )[0-9]{1,}(/|\\)[0-9]{1,}|[0-9]{1,}(/|\\)[0-9]{1,}(| )(plus|minus|times|divided(| )by)(| )[0-9]{1.}(| )[0-9]{1,}(/|\\)[0-9]{1,}"
 
-FRACTIONS=`stat --print=%s Fractions.txt`
-
-if [ $FRACTIONS -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Fractions.txt"
-  export CATNAME="Fractions"
-  $CATEGORIZE
-fi
 
-rm Fractions.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Fractions\n"
+  fi
+
+  FRACTIONS="$(egrep -i "$KEYWORDS_FRACTIONS" newpages.txt)"
+
+  if [ "$FRACTIONS" != "" ];
+  then
+    printf "%s" "$FRACTIONS" > Fractions.txt
+    export CATFILE="Fractions.txt"
+    export CATNAME="Fractions"
+    $CATEGORIZE
+    rm Fractions.txt
+    unset FRACTIONS
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Fractions\n"
+  fi
+
+fi
